// frontend/web/src/services/API/proserve-wb-provisioning-api/models/GetUserProfileResponse.ts
export interface GetUserProfileResponse {
    /**
     * Preferred region of the user as defined in the settings.
     * @type {string}
     * @memberof GetUserProfileResponse
     */
    preferredRegion?: string;
    /**
     * Enabled regions for VEW product provisioning.
     * @type {Array<string>}
     * @memberof GetUserProfileResponse
     */
    enabledRegions?: Array<string>;
    /**
     * Enabled features for the user.
     * @type {Array<GetUserProfileResponseFeature>}
     * @memberof GetUserProfileResponse
     */
    enabledFeatures?: Array<GetUserProfileResponseFeature>;
    /**
     * Preferred network of the user as defined in the settings.
     * @type {string}
     * @memberof GetUserProfileResponse
     */
    preferredNetwork?: string;
    /**
     * VEW UI Application Version.
     * @type {string}
     * @memberof GetUserProfileResponse
     */
    applicationVersion?: string;
    /**
     * VEW UI Application Version(Frontend).
     * @type {string}
     * @memberof GetUserProfileResponse
     */
    applicationVersionFrontend?: string;
    /**
     * VEW UI Application Version(Backend).
     * @type {string}
     * @memberof GetUserProfileResponse
     */
    applicationVersionBackend?: string;
    /**
     * Preferred maintenance window
     * @type {Array<MaintenanceWindow>}
     * @memberof GetUserProfileResponse
     */
    preferredMaintenanceWindows?: Array<MaintenanceWindow>;
}