// frontend/web/src/services/API/proserve-wb-provisioning-api/models/GetUserProfileResponseFeature.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Product Provisioning API
 * API in charge of provisioning VEW products
 *
 * The version of the OpenAPI document: 2024-08-12T12:00:00Z
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface GetUserProfileResponseFeature
 */
export interface GetUserProfileResponseFeature {
    /**
     * Version of the feature
     * @type {string}
     * @memberof GetUserProfileResponseFeature
     */
    version?: string;
    /**
     * Feature name
     * @type {string}
     * @memberof GetUserProfileResponseFeature
     */
    feature?: string;
    /**
     * Enabled/disabled flag
     * @type {boolean}
     * @memberof GetUserProfileResponseFeature
     */
    enabled?: boolean;
}

/**
 * Check if a given object implements the GetUserProfileResponseFeature interface.
 */
export function instanceOfGetUserProfileResponseFeature(value: object): boolean {
    let isInstance = true;

    return isInstance;
}

export function GetUserProfileResponseFeatureFromJSON(json: any): GetUserProfileResponseFeature {
    return GetUserProfileResponseFeatureFromJSONTyped(json, false);
}

export function GetUserProfileResponseFeatureFromJSONTyped(json: any, ignoreDiscriminator: boolean): GetUserProfileResponseFeature {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'version': !exists(json, 'version') ? undefined : json['version'],
        'feature': !exists(json, 'feature') ? undefined : json['feature'],
        'enabled': !exists(json, 'enabled') ? undefined : json['enabled'],
    };
}

export function GetUserProfileResponseFeatureToJSON(value?: GetUserProfileResponseFeature | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'version': value.version,
        'feature': value.feature,
        'enabled': value.enabled,
    };
}