// frontend/web/src/services/API/proserve-wb-provisioning-api/models/UpdateUserProfileRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Product Provisioning API
 * API in charge of provisioning VEW products
 *
 * The version of the OpenAPI document: 2024-08-12T12:00:00Z
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { MaintenanceWindow } from './MaintenanceWindow';
import {
    MaintenanceWindowFromJSON,
    MaintenanceWindowFromJSONTyped,
    MaintenanceWindowToJSON,
} from './MaintenanceWindow';

/**
 * 
 * @export
 * @interface UpdateUserProfileRequest
 */
export interface UpdateUserProfileRequest {
    /**
     * Preferred region of the user as defined in the settings.
     * @type {string}
     * @memberof UpdateUserProfileRequest
     */
    preferredRegion?: string;
    /**
     * Preferred network of the user as defined in the settings.
     * @type {string}
     * @memberof UpdateUserProfileRequest
     */
    preferredNetwork?: string;
    /**
     * Preferred maintenance window
     * @type {Array<MaintenanceWindow>}
     * @memberof UpdateUserProfileRequest
     */
    preferredMaintenanceWindows: Array<MaintenanceWindow>;
}

/**
 * Check if a given object implements the UpdateUserProfileRequest interface.
 */
export function instanceOfUpdateUserProfileRequest(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "preferredMaintenanceWindows" in value;

    return isInstance;
}

export function UpdateUserProfileRequestFromJSON(json: any): UpdateUserProfileRequest {
    return UpdateUserProfileRequestFromJSONTyped(json, false);
}

export function UpdateUserProfileRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): UpdateUserProfileRequest {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'preferredRegion': !exists(json, 'preferredRegion') ? undefined : json['preferredRegion'],
        'preferredNetwork': !exists(json, 'preferredNetwork') ? undefined : json['preferredNetwork'],
        'preferredMaintenanceWindows': ((json['preferredMaintenanceWindows'] as Array<any>).map(MaintenanceWindowFromJSON)),
    };
}

export function UpdateUserProfileRequestToJSON(value?: UpdateUserProfileRequest | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'preferredRegion': value.preferredRegion,
        'preferredNetwork': value.preferredNetwork,
        'preferredMaintenanceWindows': ((value.preferredMaintenanceWindows as Array<any>).map(MaintenanceWindowToJSON)),
    };
}